Crate iri_s

IRI simple wrapper

This module contains a simple wrapper to work with IRIs The main goal is that we can use a simple interface to work with IRIs which could be adapted to different implementations in the future if needed.

The library provides the macro iri to create IriS from strings.

Modules

error

Macros

iri

Generates an IriS from a string literal.

iri_once

This macro creates a static IriS variable that is initialized once and can be accessed globally.

static_once

This macro creates a static variable that is initialized once andm can be accessed globally.

Structs

IriS

Enums

Iri

An IRI that can be either a raw String or a parsed IriS

Traits

MimeType

Types that implement this trait can inform about their MIME-TYPE More information about MIME types: https://www.iana.org/assignments/media-types/media-types.xhtml