Crate iri_s

IRI simple wrapper

This module contains a simple wrapper to work with IRIs The main goal is that we can use a simple interface to work with IRIs which could be adapted to different implementations in the future if needed.

The library provides the macro iri to create IRIs from strings.

Re-exports

pub use iris::*;

pub use iris_error::*;

Modules

iris

iris_error

Macros

iri